+++ editors/code/CHANGELOG.md
@@ -1,1 +1,7 @@
-# Changelog
+# v-analyzer VS Code Extension Changelog
+
+## [0.0.1] - 03.07.2023
+
+First public release.
+
+Please note that this is a beta version, so it may contain any bugs.